Add safety check to delete

+++ src/cmds.py
@@ -210,12 +210,23 @@ def cmd_list(config, gh, org, args):
 
 cmd_delete_parser = customargs.ArgumentParser(prog='delete')
 cmd_delete_parser.add_argument('name', help="Which repository to delete")
+cmd_delete_parser.add_argument('--force', help="Forces the deletion of repositories with more than the default commit count", action="store_true")
 
 def cmd_delete(config, gh, org, args):
     repos = get_challenge_repos(config, org)
 
     for r in repos:
         if r.name == args.name:
+            total_commits = 0
+            for c in r.get_contributors():
+                total_commits += c.contributions
+
+            # the repository has been modified
+            if total_commits > 2 and not args.force:
+                log.warning("Refusing to delete repository with more commits than the default")
+                log.warning("Specify the --force flag to confirm the deletion")
+                return
+
             delete_repo(r)
             return
